tests/wpt/tests/css/css-gaps/tentative/parsing/gap-decorations-style-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Gap Decoration: column-rule-style getComputedStyle()</title>
<link rel="help" href="https://kbabbitt.github.io/css-gap-decorations/pr-11115/Overview.html#column-row-rule-style">
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<meta name="assert" content="column-rule-style computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("column-rule-style", "none");
test_computed_value("column-rule-style", "hidden");
test_computed_value("column-rule-style", "dotted");
test_computed_value("column-rule-style", "dashed");
test_computed_value("column-rule-style", "solid");
test_computed_value("column-rule-style", "double");
test_computed_value("column-rule-style", "groove");
test_computed_value("column-rule-style", "ridge");
test_computed_value("column-rule-style", "inset");

// TODO(crbug.com/357648037): Add tests for multiple values when parsing is implemented.

</script>
</body>
</html>
